feat: change frame from u32 -> f32, this helps to make the animation smoother

+++ crates/core/src/layer/opacity.rs
@@ -11,10 +11,10 @@ impl AnimatedExt for OpacityHierarchy {
     type Target = f32;
 
     fn initial_value(&self) -> Self::Target {
-        self.value(0)
+        self.value(0.0)
     }
 
-    fn value(&self, frame: u32) -> Self::Target {
+    fn value(&self, frame: f32) -> Self::Target {
         self.stack
             .iter()
             .fold(1.0, |current, item| current * item.value(frame) / 100.0)
